Tweaked some osx specific rendering setup.
There are some timing issues in initialization of the
renderer in the main thread vs. using it in the primary
TaskMaster thread. Windows didn't have these problems,
at least in Debug builds, but Clang certainly does.

I need to revise how we initialize the renderer and make
sure that happens on the main TaskMaster thread, once
that's done, I'll pick up the osx work again.